Rename http_vpc file & fix requires
